# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateFarmRequest:
    properties:
      city:
        type: string
      location:
        type: string
      name:
        maxLength: 150
        minLength: 2
        type: string
      owner_name:
        maxLength: 150
        minLength: 2
        type: string
      state:
        type: string
      total_area:
        type: number
    required:
    - city
    - name
    - owner_name
    - state
    - total_area
    type: object
  dto.CreateFieldRequest:
    properties:
      area:
        type: number
      farm_id:
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      soil_type:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/entities.FieldStatus'
        enum:
        - active
        - inactive
        - fallow
    required:
    - area
    - farm_id
    - name
    type: object
...
  dto.FarmResponse:
    properties:
      city:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      owner_name:
        type: string
      state:
        type: string
      total_area:
        type: number
      updated_at:
        type: string
    type: object
  dto.FieldResponse:
    properties:
      area:
        type: number
      created_at:
        type: string
      created_by:
        type: integer
      farm_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      soil_type:
        type: string
      status:
        $ref: '#/definitions/entities.FieldStatus'
      updated_at:
        type: string
    type: object
...
  dto.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.UpdateFarmRequest:
    properties:
      city:
        type: string
      location:
        type: string
      name:
        maxLength: 150
        minLength: 2
        type: string
      owner_name:
        maxLength: 150
        minLength: 2
        type: string
      state:
        type: string
      total_area:
        type: number
    required:
    - city
    - name
    - owner_name
    - state
    - total_area
    type: object
  dto.UpdateFieldRequest:
    properties:
      area:
        type: number
      farm_id:
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      soil_type:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/entities.FieldStatus'
        enum:
        - active
        - inactive
        - fallow
    required:
    - area
    - farm_id
    - name
    - status
    type: object
...
  entities.FieldStatus:
    enum:
    - active
    - inactive
    - fallow
    type: string
    x-enum-comments:
      FieldStatusFallow: pousio
    x-enum-descriptions:
    - ""
    - ""
    - pousio
    x-enum-varnames:
    - FieldStatusActive
    - FieldStatusInactive
    - FieldStatusFallow
...
  utils.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API REST de gestão agrícola com autenticação JWT
  title: AgroControl API
  version: "1.0"
paths:
...
  /api/v1/farms:
    get:
      description: Retorna fazendas paginadas. Admin vê todas, outros veem apenas
        as suas.
      parameters:
      - description: Página (padrão 1)
        in: query
        name: page
        type: integer
      - description: Itens por página (padrão 10)
        in: query
        name: limit
        type: integer
      - description: Busca por nome
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar fazendas
      tags:
      - farms
    post:
      consumes:
      - application/json
      description: Cria uma nova fazenda. Requer role manager ou admin.
      parameters:
      - description: Dados da fazenda
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFarmRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.FarmResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar fazenda
      tags:
      - farms
  /api/v1/farms/{id}:
    delete:
      description: Exclui uma fazenda. Requer role admin.
      parameters:
      - description: ID da fazenda
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Excluir fazenda
      tags:
      - farms
    get:
      description: Retorna os dados de uma fazenda específica
      parameters:
      - description: ID da fazenda
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FarmResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buscar fazenda por ID
      tags:
      - farms
    put:
      consumes:
      - application/json
      description: Atualiza os dados de uma fazenda. Requer role manager ou admin.
      parameters:
      - description: ID da fazenda
        in: path
        name: id
        required: true
        type: integer
      - description: Dados atualizados
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateFarmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FarmResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar fazenda
      tags:
      - farms
  /api/v1/farms/{id}/fields:
    get:
      parameters:
      - description: ID da fazenda
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.FieldResponse'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar talhões de uma fazenda
      tags:
      - fields
  /api/v1/fields:
    get:
      description: Retorna talhões paginados. Admin vê todos, outros veem apenas os
        seus.
      parameters:
      - description: Página
        in: query
        name: page
        type: integer
      - description: Itens por página
        in: query
        name: limit
        type: integer
      - description: Busca por nome
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: Listar talhões
      tags:
      - fields
    post:
      consumes:
      - application/json
      description: Cria um novo talhão em uma fazenda. Valida se o usuário tem acesso
        à fazenda.
      parameters:
      - description: Dados do talhão
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFieldRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.FieldResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Fazenda não encontrada
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar talhão
      tags:
      - fields
  /api/v1/fields/{id}:
    delete:
      parameters:
      - description: ID do talhão
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Excluir talhão
      tags:
      - fields
    get:
      parameters:
      - description: ID do talhão
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FieldResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buscar talhão por ID
      tags:
      - fields
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do talhão
        in: path
        name: id
        required: true
        type: integer
      - description: Dados atualizados
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateFieldRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FieldResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar talhão
      tags:
      - fields
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"